requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
beautifulsoup4==4.12.3
aiohttp==3.9.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.4
python-multipart==0.0.6
aiohttp[speedups]
fake-useragent==1.4.0
geopy==2.4.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
asyncio==3.4.3
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3